공지사항, FAQ 수정 및 관리자 비밀번호 초기화 수정

+++ src/main/java/kr/co/infovine/dkmm/db/model/board/TBoard.java
@@ -16,7 +16,7 @@ public class TBoard {
      *
      * @mbg.generated Wed Dec 07 09:59:40 KST 2022
      */
-    private Integer boardSeq;
+    private int boardSeq;
 
     /**
      *
@@ -52,7 +52,7 @@ public class TBoard {
      *
      * @mbg.generated Wed Dec 07 09:59:40 KST 2022
      */
-    private Long viewCnt;
+    private int viewCnt;
 
     /**
      *
@@ -70,7 +70,7 @@ public class TBoard {
      *
      * @mbg.generated Wed Dec 07 09:59:40 KST 2022
      */
-    private Long insSeq;
+    private int insSeq;
 
     /**
      *
@@ -89,7 +89,7 @@ public class TBoard {
      *
      * @mbg.generated Wed Dec 07 09:59:40 KST 2022
      */
-    private Long uptSeq;
+    private int uptSeq;
 
     /**
      *
